+++ packages/snap/src/core/handlers/onKeyringRequest/Keyring.ts
@@ -53,7 +53,11 @@ import { SolanaWalletRequestStruct } from '../../services/wallet/structs';
 import type { WalletService } from '../../services/wallet/WalletService';
 import { deriveSolanaKeypair } from '../../utils/deriveSolanaKeypair';
 import { getLowestUnusedIndex } from '../../utils/getLowestUnusedIndex';
-import { listEntropySources } from '../../utils/interface';
+import {
+  endTrace,
+  listEntropySources,
+  startTrace,
+} from '../../utils/interface';
 import { createPrefixedLogger, type ILogger } from '../../utils/logger';
 import {
   DeleteAccountStruct,
@@ -87,6 +91,8 @@ export class SolanaKeyring implements Keyring {
   readonly #keyringAccountMonitor: KeyringAccountMonitor;
 
   readonly #nameResolutionService: NameResolutionService;
+
+  readonly #traceName: string = 'Create Solana Account';
 
   constructor({
     state,
@@ -217,6 +223,8 @@ export class SolanaKeyring implements Keyring {
     const id = globalThis.crypto.randomUUID();
 
     try {
+      await startTrace(this.#traceName);
+
       const accounts = await this.listAccounts();
 
       const entropySource =
@@ -347,6 +355,8 @@ export class SolanaKeyring implements Keyring {
           },
         },
       });
+
+      await endTrace(this.#traceName);
 
       return keyringAccount;
     } catch (error: any) {