+++ packages/snap/src/core/services/analytics/AnalyticsService.ts
@@ -46,10 +46,8 @@ export class AnalyticsService {
       await this.#trackEvent('Transaction Added', {
         message: 'Snap transaction added',
         origin: metadata.origin,
-        account_id: account.id,
-        account_address: account.address,
         account_type: account.type,
-        chain_id: metadata.scope,
+        chain_id_caip: metadata.scope,
       });
     } catch (error) {
       this.#logger.warn('Error tracking event transaction added', {
@@ -69,10 +67,8 @@ export class AnalyticsService {
       await this.#trackEvent('Transaction Approved', {
         message: 'Snap transaction approved',
         origin: metadata.origin,
-        account_id: account.id,
-        account_address: account.address,
         account_type: account.type,
-        chain_id: metadata.scope,
+        chain_id_caip: metadata.scope,
       });
     } catch (error) {
       this.#logger.warn('Error tracking event transaction approved', {
@@ -93,10 +89,8 @@ export class AnalyticsService {
       await this.#trackEvent('Transaction Submitted', {
         message: 'Snap transaction submitted',
         origin: metadata.origin,
-        account_id: account.id,
-        account_address: account.address,
         account_type: account.type,
-        chain_id: metadata.scope,
+        chain_id_caip: metadata.scope,
       });
     } catch (error) {
       this.#logger.warn('Error tracking event transaction submitted', {
@@ -118,10 +112,8 @@ export class AnalyticsService {
       await this.#trackEvent('Transaction Finalized', {
         message: 'Snap transaction finalized',
         origin: metadata.origin,
-        account_id: account.id,
-        account_address: account.address,
         account_type: account.type,
-        chain_id: transaction.chain,
+        chain_id_caip: transaction.chain,
         transaction_status: transaction.status,
         transaction_type: transaction.type,
       });
@@ -144,10 +136,8 @@ export class AnalyticsService {
       await this.#trackEvent('Transaction Rejected', {
         message: 'Snap transaction rejected',
         origin: metadata.origin,
-        account_id: account.id,
-        account_address: account.address,
         account_type: account.type,
-        chain_id: metadata.scope,
+        chain_id_caip: metadata.scope,
       });
     } catch (error) {
       this.#logger.warn('Error tracking event transaction rejected', {
@@ -171,10 +161,8 @@ export class AnalyticsService {
       await this.#trackEvent('Security Alert Detected', {
         message: 'Snap security alert detected',
         origin,
-        account_id: account.id,
-        account_address: account.address,
         account_type: account.type,
-        chain_id: scope,
+        chain_id_caip: scope,
         security_alert_response: securityAlertResponse,
         security_alert_reason: securityAlertReason,
         security_alert_description: securityAlertDescription,
@@ -204,10 +192,8 @@ export class AnalyticsService {
       await this.#trackEvent('Security Scan Completed', {
         message: 'Snap security scan completed',
         origin,
-        account_id: account.id,
-        account_address: account.address,
         account_type: account.type,
-        chain_id: scope,
+        chain_id_caip: scope,
         scan_status: scanStatus,
         has_security_alerts: hasSecurityAlerts,
       });